Refactors auth screens UI
Removes animated transitions and image backgrounds to simplify authentication screens.
Replaces legacy components with new UI library elements and updates styling.
