+++ app/auth/Register.tsx
@@ -1,18 +1,14 @@
 // app/register.tsx
 import { useRouter } from "expo-router";
 import React, { useState } from "react";
-import { View, Text, TextInput, TouchableOpacity, Alert } from "react-native"; // Added Alert
-import Animated, {
-  FadeInDown,
-  FadeInUp,
-  BounceIn,
-  Layout,
-  SlideInDown,
-} from "react-native-reanimated";
+import { View, TouchableOpacity, Alert } from "react-native";
 
 import useAuthStore from "../../store/authStore";
+
+import { Button } from "~/components/ui/button";
+import { Input } from "~/components/ui/input";
+import { Text } from "~/components/ui/text";
 import { supabase } from "~/utils/supabase";
-import { Button } from "~/components/Button"; // Import the Button component
 
 export default function RegisterScreen() {
   const [email, setEmail] = useState("");
@@ -68,22 +64,22 @@ export default function RegisterScreen() {
 
   return (
     <View className="flex-1 p-6 justify-center bg-white">
-      <Animated.View entering={BounceIn.delay(200).duration(800)}>
+      <View>
         <Text className="text-3xl font-bold mb-2 text-center text-gray-700">
           Daftar Akun
         </Text>
-      </Animated.View>
+      </View>
 
-      <Animated.View entering={FadeInDown.delay(400).duration(600)}>
+      <View>
         <Text className="text-center mb-8 text-gray-600">
           Buat akun baru untuk menggunakan aplikasi
         </Text>
-      </Animated.View>
+      </View>
 
-      <Animated.View entering={SlideInDown.delay(500).duration(500)}>
+      <View>
         <View className="mb-4">
           <Text className="text-gray-700 mb-2">Email</Text>
-          <TextInput
+          <Input // Use Input component
             className="border border-gray-300 rounded-lg px-4 py-2.5"
             placeholder="Masukkan email Anda"
             keyboardType="email-address"
@@ -92,12 +88,12 @@ export default function RegisterScreen() {
             onChangeText={setEmail}
           />
         </View>
-      </Animated.View>
+      </View>
 
-      <Animated.View entering={SlideInDown.delay(600).duration(500)}>
+      <View>
         <View className="mb-4">
           <Text className="text-gray-700 mb-2">Password</Text>
-          <TextInput
+          <Input // Use Input component
             className="border border-gray-300 rounded-lg px-4 py-2.5"
             placeholder="Masukkan password Anda"
             secureTextEntry
@@ -105,12 +101,12 @@ export default function RegisterScreen() {
             onChangeText={setPassword}
           />
         </View>
-      </Animated.View>
+      </View>
 
-      <Animated.View entering={SlideInDown.delay(700).duration(500)}>
+      <View>
         <View className="mb-6">
           <Text className="text-gray-700 mb-2">Konfirmasi Password</Text>
-          <TextInput
+          <Input // Use Input component
             className="border border-gray-300 rounded-lg px-4 py-2.5"
             placeholder="Masukkan kembali password Anda"
             secureTextEntry
@@ -118,28 +114,28 @@ export default function RegisterScreen() {
             onChangeText={setConfirmPassword}
           />
         </View>
-      </Animated.View>
+      </View>
 
-      <Animated.View entering={FadeInUp.delay(800).duration(600)}>
+      <View>
         <Button
-          variant="primary"
-          size="large"
+          variant="default"
+          size="lg"
           className="mb-4"
           onPress={handleRegister}
-          loading={loading}
+          disabled={loading}
         >
-          Daftar
+          <Text>{loading ? "Loading..." : "Daftar"}</Text>
         </Button>
-      </Animated.View>
+      </View>
 
-      <Animated.View entering={FadeInUp.delay(1000).duration(600)}>
+      <View>
         <View className="flex-row justify-center mt-4">
           <Text className="text-gray-600">Sudah memiliki akun? </Text>
           <TouchableOpacity onPress={() => router.push("/auth/Login")}>
             <Text className="text-gray-700 font-semibold">Masuk</Text>
           </TouchableOpacity>
         </View>
-      </Animated.View>
+      </View>
     </View>
   );
 }
